requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
email-validator
sqlalchemy
psycopg[binary]
langchain
langchain-community
pinecone
python-dotenv
langchain-classic
